Power(x,n)

+++ Step 7 - Recursion Patternwise/Lecture 1 - Get a Strong Hold/Pow_x_n.cpp
@@ -30,7 +30,7 @@ int main()
   double x;
   int n;
   cin >> x >> n;
-  int result = myPow(x, n);
+  double result = myPow(x, n);
   cout << result << endl;
   return 0;
 }
